refactor(core): Relocate logic to finalize modularization
- Moves `calculateBalance` and `handleTransactionAction` out of `core/main.js`.
- Ensures `core/main.js` acts as a pure initialization script by removing all non-setup related functions.
- Improves code cohesion by placing transaction logic (`calculateBalance`) and UI event handling (`handleTransactionAction`) directly where they are used.
- Updates all necessary import paths across the application to reflect the new locations of the moved functions.
