+++ core/main.js
@@ -7,13 +7,13 @@ import { app } from './state.js';
 import { initModal } from '../ui/modal.js';
 import { initFab } from '../ui/fab.js';
 import { initNavigation } from '../ui/navigation.js';
-import { showAlert, showConfirm } from '../ui/notifications.js';
-import { initRenderer, render } from '../ui/renderer.js';
+import { showConfirm } from '../ui/notifications.js';
+import { render } from '../ui/renderer.js';
 
 // Import Feature Modules
-import { initActions, deleteTransaction, deletePayment } from '../features/actions.js';
+import { initActions } from '../features/actions.js';
 import { initPersonModals } from '../features/persons/person-modals.js';
-import { initTransactionModals, showEditTransactionModal, showPaymentModal, showTransactionDetails } from '../features/transactions/transaction-modals.js';
+import { initTransactionModals } from '../features/transactions/transaction-modals.js';
 import { exportData, handleImport } from '../features/import-export/data-service.js';
 
 
@@ -28,17 +28,10 @@ export async function init() {
     await db.init();
     await loadData();
 
-    // Pass dependencies to the renderer module
-    initRenderer({
-        calculateBalance,
-        handleTransactionAction,
-    });
-
     // Initialize feature modules
     initActions({
         loadData,
         render,
-        calculateBalance,
     });
 
     initPersonModals({
@@ -47,8 +40,6 @@ export async function init() {
     });
 
     initTransactionModals({
-        calculateBalance,
-        deletePayment,
         loadData,
         render,
     });
@@ -90,50 +81,7 @@ function setupEventListeners() {
 }
 
 // =================================================================================================
-// DATA MANIPULATION & ACTIONS
-// =================================================================================================
-
-/**
- * Calculates the remaining balance for a transaction.
- * @param {object} transaction - The transaction object.
- * @returns {number} The remaining balance in cents.
- */
-function calculateBalance(transaction) {
-    const totalPaid = transaction.payments?.reduce((sum, p) => sum + p.amount, 0) || 0;
-    return transaction.amount - totalPaid;
-}
-
-/**
- * Handles clicks on action buttons within a transaction card.
- * @param {Event} e - The click event.
- */
-function handleTransactionAction(e) {
-    const action = e.target.dataset.action;
-    const id = e.target.dataset.id;
-
-    // For delete, the transaction object may no longer exist.
-    // For other actions, we find it.
-    const transaction = app.transactions.find(t => t.id === id);
-
-    switch (action) {
-        case 'payment':
-            showPaymentModal(transaction);
-            break;
-        case 'details':
-            showTransactionDetails(transaction);
-            break;
-        case 'edit':
-            showEditTransactionModal(transaction);
-            break;
-        case 'delete':
-            // Call the imported delete function
-            deleteTransaction(id);
-            break;
-    }
-}
-
-// =================================================================================================
-// UTILITIES & SERVICE WORKER
+// SERVICE WORKER
 // =================================================================================================
 
 /**
